cli: unify script to package; package CLI gains --contains/--all and count --json/--output; fix NFPA arrow symbols; add Windows CI matrix; add docs/CLI.md; update docs
